Changed the on data function
Re-wrote the socket.on('data') function. Now we wait untill we have a full line received, this way we dont end up getting 1 command in 2 packets (resulting in an extra, mostly empty, command, returning an ACK).

Now we are also able to extract the command and its parameters better from the line, and put them in a switch-case.
The switch-case is mostly here to replace the .startsWith() method. It also looks cleaner.