Fix incorrect invocation of isMounted
The second argument of isMounted should be 'true' if the path given
is a device and 'false 'if the path is mountpoint. Here we are definitely
dealing with a mountpoint. A bad side effect of this is that when the test
fails, which it always will, is that it tries to rm -rf everything under
the given path. This could possibly wipe out a music collection...