Add files via upload
This fix takes care of the case when Repeat mode is on. If the current track is the last track of the queue the next track will be the first track. It has been tested also with the other modes. Used in combination with other modes Random and Consume override Repeat. Repeat + Random = Random; Repeat + Consume = Consume; Random + Repeat + Consume = Repeat + Consume, but this combination is still buggy (same bugs as before).