Avoid unsuitable characters in mountpoints (#1010)
* Remove variable unused in this function

* Use a more robust scheme for local mountpoint names

Instead of escaping spaces, use an underscore in place of all characters that
get encoded by setmntent(7) in /etc/mtab entries. Use this method consistently
in all functions that have to mount or unmount a share.

* Use mountidraw instead of recomputing its value

* Found another mountpoint construction to fix

* Avoid trying to operate on undefined values

When a share is deleted, this happens asynchronously. It can happen that we
try to look up a share that has recently been deleted and get undefined values
from the lookup. Operating on undefined values can cause a crash, so try
to detect this and return early.