Fix incorrect invocation of isMounted
The second argument of isMounted should be 'true' if the path given
is a device and 'false' if the path is mountpoint. Here we are definitely
dealing with a mountpoint so using 'true' causes the isMounted test to fail.
A bad side effect of this is that when the test fails, which it always will,
the code then tries to rm -rf everything under the given path rather than
trying to umount the path. Without readonly mounts, this could possibly
wipe out a music collection...