Deleteshare fixes (#1020)
* Fix incorrect invocation of isMounted

The second argument of isMounted should be 'true' if the path given
is a device and 'false' if the path is mountpoint. Here we are definitely
dealing with a mountpoint so using 'true' causes the isMounted test to fail.
A bad side effect of this is that when the test fails, which it always will,
the code then tries to rm -rf everything under the given path rather than
trying to umount the path. Without readonly mounts, this could possibly
wipe out a music collection...

* Change how deleteShare unmounts the share to match editShare

The existing code reimplements what mountutils.umount() does, but with more
error reporting. The way it calls sudo relies on the volumio user password
never changing. Mountutils can do what we need; editShare uses it already.

* Fix indentation

* Only remove the share key on successful deletion

By unconditionally removing the share's key from the configuration, users lose
the ability to interact with the share through the UI. If umounting fails,
leave the share visible so that they can try removing it again at a later time.

* Update the logged error message

If the share could not be removed, it should still exist in the on-disk
configuration file since we no longer unconditionally delete the config key.
So volumio should attempt to mount it again after a reboot.