Handle mount errors where the reason is undefined
There are cases where a NAS mount fails and no reason is returned by mountShare.
An example case is when the user specifies an incorrect path on the remote host.
When this occurs, data.reason is undefined and in the current code this leads
to the share being saved in the configuration data, but not appearing (as a
not-mounted drive) in the list of drives. The user gets no feedback, either.
Fix by adding a resolve() for that case and pushing a message to the user.