Rework parsing of smbtree output
The old code mostly worked, but share names with spaces were not handled;
'my stuff' ended up as 'my'. The main fix here is splitting the output lines
using the field delimiter (tab). Also it seemed clearer to collate NAS names
as keys in an object with arrays of share names as the values, at the (small)
cost of converting the object into the format Volumio2-UI is expecting.