Improve discoverShares() (#1029)
* indentation

* Stop smbtree asking for a password

* Use a broadcast query instead of being limited to the local master browser

* Rework parsing of smbtree output

The old code mostly worked, but share names with spaces were not handled;
'my stuff' ended up as 'my'. The main fix here is splitting the output lines
using the field delimiter (tab). Also it seemed clearer to collate NAS names
as keys in an object with arrays of share names as the values, at the (small)
cost of converting the object into the format Volumio2-UI is expecting.

* Change variable name to one that is harder to type incorrectly.

Fixes an undefined variable exception that was causing an empty list to be
returned.

* Fix handling of share names padded to 16 characters with trailing spaces

The share name defined in smb.conf does not include the spaces,
so we have to strip them or we will not be able to mount the share.
Shares that _do_ have trailing spaces in their definitions on the server
are not supported by this code.