More robust mounting (#1057)
* Improve handling of paths with 'tricky' characters

The existing mountShare code takes some trouble to escape any spaces in
a NAS share name or NFS path. But this ignores other characters, such as '('.
Under the hood the mount is done as a shell exec call to 'mount'; with such
characters (which the shell treats specially) in the path, the mount fails.
Related to this, NAS aliases (used to construct mountpoint paths) could also
have characters that cause problems, such as (, " or '.

To fix all this, first we have to fix linux-mountutils, see the patch below.
This allows us to ditch the space-escaping of remote paths and ensures that
'raw' mountpoint paths we pass to linux-mountutils are handled correctly.
We also need to take the same care with quoting when shell execs are performed
within this plugin, such as in getMountSize().

  --- node_modules/linux-mountutils/mountutils.js.orig    2017-03-06 23:39:28.132120603 +1100
  +++ node_modules/linux-mountutils/mountutils.js 2017-03-07 00:06:41.765539915 +1100
  @@ -6,6 +6,23 @@
    *
    */

  +/*
  + * Hide characters the shell is sensitive to inside single quotes.
  + * Shell quoting rules require this to be done by:
  + * splitting on any single quotes, wrapping each piece in single quotes
  + * and joining the pieces with escaped single quotes.
  + */
  +var quotepath = function(path) {
  +    var output = '';
  +    var pieces = path.split("'");
  +    var n = pieces.length;
  +    for (var i=0; i<n; i++) {
  +        output = output + "'" + pieces[i] + "'";
  +        if (i< (n-1)) output = output + "\\'";
  +    }
  +    return output;
  +}
  +
   /**
    * isMounted
    *
  @@ -94,6 +111,10 @@
       return;
     }

  +  // properly quote paths so the shell doesn't explode
  +  var qdev  = quotepath(dev);
  +  var qpath = quotepath(path);
  +
     // Build the command line
     var cmd = (options.noSudo?"":
         (options.sudoPath?options.sudoPath:"/usr/bin/sudo")+" ") +
  @@ -101,7 +122,7 @@
         (options.readonly?"-r ":"") +
         (options.fstype?"-t " + options.fstype + " ":"") +
         (options.fsopts?"-o " + options.fsopts + " ":"") +
  -      dev + " " + path;
  +      qdev + " " + qpath;

     // Let's do it!
     var mountProc = exec(cmd, function(error, stdout, stderr) {
  @@ -137,10 +158,11 @@
       return;
     }

  +  var qpath = quotepath(path);
     // Build the command line
     var cmd = (options.noSudo?"":
         (options.sudoPath?options.sudoPath:"/usr/bin/sudo")+" ") +
  -      (options.umountPath?options.umountPath:"/bin/umount") + " " + path;
  +      (options.umountPath?options.umountPath:"/bin/umount") + " " + qpath;

     // Let's do it!
     var umountProc = exec(cmd, function(error, stdout, stderr) {

* Make sure unexpected error messages reach the caller

The split on the string 'mount error' throws away some preamble text
that clutters up the error message. But if that string does not appear
in the error message, the code as it stands throws away the entire message.

* Fix indentation