Fix issue #1161
This fixes issue #1161

It will use any correcly names images inside the album folder as album art and display it on the album list page.

On boot it will run the albumList function and store the result in cache.  On first request to view albums it will read the cache album list file and copy any matching art files in the album file to the volumio albumart folder and only download art for albums that have no album art file.

The cached result will be updated after a library update.

Images in the browser cache should be cleared before first running with this update.

Subsequent requests to view albums will be loaded from cache by volumio and the images will usually be loaded from the browser cache.