Improve user input hygiene in logsubmit.js (#1285)
* Use single-quoting around the message included with the log.

If we don't then things like parentheses will be interpreted by the shell.
See for example #1237.

* Another case needing single-quoting

* Lay out construction of curl command so it is readable

* Invert the quoting when constructing the curl command

Single-quoting the (only partly parsed) form data ensures the shell does not
try to expand it. Without this certain input descriptions can lead to the curl
call crashing and the bug report appearing to take forever to be submitted.