Revisit quoting of description when submitting logs
First attempt at this did not handle the case of a single-quote within the
description.  The solution is to break the string on any single-quote,
quote the pieces either side of it and escape the original single-quote.
This must be done both when calling logsumbit.js and when logsumit.js
needs to expose the description string to the shell.