More logsubmit work (#1292)
* Revisit quoting of description when submitting logs

First attempt at this did not handle the case of a single-quote within the
description.  The solution is to break the string on any single-quote,
quote the pieces either side of it and escape the original single-quote.
This must be done both when calling logsumbit.js and when logsumit.js
needs to expose the description string to the shell.

* Use correct variable name in catch() clauses

* Clean input data before attempting to parse

In issue #1237 I noticed the browser dev tools were highlighting an error:

  SyntaxError: JSON.parse: unexpected character at line 1 column 1
  of the JSON data

A bit of print-statement debugging showed that the json string being emitted
from logsubmit.js was being contaminated somehow. An example (line-wraps added)

  description: 'simple test 3'
  /usr/bin/curl -X POST -H 'Content-Type: multipart/form-data' \
    -F 'logFile=@/tmp/logondemand' -F desc='simple test 3' \
    -F 'variant=volumio' 'http://logs.volumio.org:7171/logs/v1'
  {"status":"OK","link":"http://logs.volumio.org/volumio/Ig63ngo.html"}

Until the cause of this can be tracked down, remove any unwanted text before
or after the json data string before handing it to JSON.parse().