Properly escape strings fed to the mpd 'find' command.
Consider this URI beng passed to listAlbumSongs():
  albums://Kalafina/THE%20BEST%20%22Blue%22
which decodes to this:
  albums://Kalafina/THE BEST "Blue"

When the current code constructs an mpd 'find' command, it produces this:
  find album "THE BEST "Blue"" albumartist "Kalafina"
which causes a failure in MpdClient.receive()
  Error:  [5@0] {} Space expected after closing '"'

What we want to send to mpd is:
  find album "THE BEST \"Blue\"" albumartist "Kalafina"

To get this result, we must escape any double quotes within the strings
that are concatenated to form the 'find' command.
No attempt is made to handle the pathological case of the input string
containing '\"'.