Make the PATH that is in use unambiguous
When volumio.sh is called by the volumio user from the command line,
it gets one PATH but when it is called by shelling out from the volumio
application, it gets quite another. This is because when shelling out
it may run as user 'volumio' but it inherits its PATH from 'root',
since 'root' starts volumio.service.
The right thing to do (tm) is just be explicit about what PATH is needed.