Save logsubmit logs locally (#1473)
* Open log file only after argument parsing

* If we can't upload the log, store it

Sometimes the log upload fails, but it would be nice to keep the
data around so it can be sent by other means. /var/tmp should
persist across reboots, so move the log there.

* Add support for storing log locally, rather than uploading.

In cases where the user knows that an upload attempt will fail
(eg network issues), trying and failing will just be confusing.
To address this, support a second argument to explicitly save
the bug report to local storage.

* Add 'logdump' option to save bug report logs locally

* Drop use of /tmp/logfields

All this file was doing was storing the description line.
Now that we can reliably pass the description to logsubmit.js,
we no longer need it.

* Ensure a parseable description line

The now-removed logfilds file had one useful element -
a line that made it easier to extract the description
input by the user. Restore the line to its parseable form.

* Make the PATH that is in use unambiguous

When volumio.sh is called by the volumio user from the command line,
it gets one PATH but when it is called by shelling out from the volumio
application, it gets quite another. This is because when shelling out
it may run as user 'volumio' but it inherits its PATH from 'root',
since 'root' starts volumio.service.
The right thing to do (tm) is just be explicit about what PATH is needed.

* Whitespace

* Add section headings for each command attempted

* Whitespace