Enforce leading / on NFS path supplied by user
NFS mounts require an absolute path to the exported directory on the server.
The volumio UI does not encourage this, as the default is mouting via CIFS.
To avoid troubling the user with these details, fix up the path before the
configuration is used or saved to disk. See Volumio2-UI#511.