Search and play non-VA albums that contain multiple artists
Often when index.js checks to see if we are playing a compilation, but we are not, it then relies on the 'artist' field of tracks to decide if they are part of the same album to display or play. This patch makes these occasions rely instead on 'albumartist', which should hopefully correctly play more albums.

For example, if an album by 'Sting' contains a track with the artist listed as 'Sting feat. Beethoven', when you find that album under the artist view, it would previously not list the track featuring Beethoven. After this patch, it would list all the tracks on the album.