Pass all the arguments to pull.sh
The pull command is supposed to support the ability to pull a
specific branch, either from the official Volumio repo or from a
custom repo.
Is seems like this was either not working or a regression has been
introduced at some point.
This commit fixes the behaviour by passing to the pull.sh script
all the parameters passed to the pull() funciton in volumio.sh.