Myvolumio plugin loader split (#1700)
* Split MyVolumio Plugin Management from PluginManager

Since we need to be able to have a quite different approach on how
we handle MyVolumio plugins, it's better to split this out of
the original PluginManager.

In this commit:
* Create a myvolumio-pluginmanager.js file, which handles load/start
  activities for MyVolumio related plugins
* Make sure this new file is compliant with substandard
* Encapsulate myvolumio-pluginmanager within the orignal PluginManager,
  so that the external interface stays unchanged

* Ensure compatibility for getPluginCategories/Names

These methods in PluginManager are expected to deliver the list of
all the plugin names / categories, including the ones handled via
MyVolumio. Make sure to include them.

Additionally, print a useful error message if a package.json file
cannot be read.

* Pass configManager to child module

Instead of having to require the configManager from myvolumio's
plugin manager, pass it over as function parameter.

* Remove myvolumio-pluginmanager

This code has now been moved to a separate module, it does not need
to be distributed from this repository anymore.