Delegate to MyVolumio the handling of additional conf (#1776)
Instead of directly referencing a MyVolumio plugin from the main "My
Music" plugin, delegate the task of fetching in additional UIConfig from
any plugin to the PluginManager. This will make it easier in the future
to change the way MyVolumio plugins behave without having to change code
in the core plugins.