Check before overwriting .name field of a queue item.
Play queue items contain a valid .name field but not necessarily
a valid .title field. One instance where this happens when a webradio
is saved to a playlist and the playlist is added to the play queue.
Check .title is defined before overwriting the .name field.
Otherwise, the item will be shown in the queue with a name of "undefined".