RemoveFromFavourites now checks for the correspondent method on plugin. If found the method is called. Otherwise the Common Method is called to handle the request.
Enhanced to push an update of the browse library.

WHAT SHALL BE DONE: Check if the favourite mechanism through the common method has been broken (couldn't do right now)

WHAT SHALL BE ANALYZED AND DONE IN THE FUTURE: When the Browse is pushed the user can click the back button. This loads the page where the removed song is present. Check why this happens.