clear repeatSingle if repeat is disabled
repeat single does not work alone, since its nested inside repeat checks and coupled with it in other checks, so it should be set false when repeat is disabled. Currently "pushState" can emit "repeatSingle == true", but if repeat is false, we are not repeating single track, so the state is wrong.