Ffongmdsp device volume db display (#2148)
* Send device volume (in dB units) to Volumio client as part of the
'pushState' command. This field, named 'deviceVolume' will present
alongside the percentage-volume value 'volume' field of the data
argument of 'pushState'.

* Do not assume alsavolume() always return a Promise. It could be
null. It means caller should not set up a callback-function on
this operation.

* Renamed expected Volume object field 'deviceVol' to 'dbVolume'
as suggested by Volumio. Refer to corresponding change at
'inputs' plugin.

* Revert "Do not assume alsavolume() always return a Promise. It could be"

This reverts commit 08b37a7282d05aa4a5383743f7b115b4e03d8021.

By request of Volumio, it is only necessary when the corresponding
change in 'inputs' plugin is applied. (Overriding the alsavolume())

* Renamed the state field 'deviceVolume' to 'dbVolume' by request
of Volumio.

* Naming consistency

Co-authored-by: Frankie Fong <frankie@minidsp.com>